Reworked bundle submission component for Reusable
