+++ forms-flow-web/src/actions/actionConstants.js
@@ -180,6 +180,7 @@ const ACTION_CONSTANTS = {
   BUNDLE_FORM_LIST_FORM_SEARCH: "BUNDLE_BPM_FORM_SEARCH",
   BUNDLE_FORM_SUBMISSION: "BUNDLE_FORM_SUBMISSION",
   BUNDLE_LOADING: "BUNDLE_LOADING",
+  BUNDLE_SUBMIT_LOADING: "BUNDLE_SUBMIT_LOADING",
 };
 
 export default ACTION_CONSTANTS;
