+++ forms-flow-web/src/components/Bundle/item/BundleSubmit.js
@@ -1,166 +1,87 @@
-import React, { useState, useEffect, useRef } from "react";
-import { Link } from "react-router-dom";
+import React, { useEffect } from "react";
+import BundleSubmission from "./BundleSubmission";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchFormById } from "../../../apiManager/services/bpmFormServices";
-import Loading from "../../../containers/Loading";
-import { Form, Errors } from "react-formio";
-import Stepper from "@material-ui/core/Stepper";
-import Step from "@material-ui/core/Step";
-import StepLabel from "@material-ui/core/StepLabel";
-import { MULTITENANCY_ENABLED, STAFF_DESIGNER } from "../../../constants/constants";
-import { Formio } from "formiojs";
-import { toast } from "react-toastify";
-import {
-  setBundleSelectedForms,
-  setBundleSubmissionData,
-} from "../../../actions/bundleActions";
-import { formioPostSubmission } from "../../../apiManager/services/FormServices";
-import LoadingOverlay from "react-loading-overlay";
-import { push } from "connected-react-router";
-import {
-  clearFormError,
-  setFormFailureErrorData,
-  setFormSubmissionError,
-} from "../../../actions/formActions";
-import SubmissionError from "../../../containers/SubmissionError";
+import { setBundleLoading, setBundleSelectedForms, setBundleSubmissionData, setBundleSubmitLoading } from "../../../actions/bundleActions";
+import { clearFormError, clearSubmissionError, setFormFailureErrorData, setFormSubmissionError } from "../../../actions/formActions";
+import { getFormProcesses } from "../../../apiManager/services/processServices";
 import { executeRule } from "../../../apiManager/services/bundleServices";
+import { Link, useParams } from "react-router-dom";
+import { MULTITENANCY_ENABLED, STAFF_DESIGNER } from "../../../constants/constants";
+import Loading from "../../../containers/Loading";
+import { Errors } from "react-formio/lib/components";
+import SubmissionError from "../../../containers/SubmissionError";
+import { formioPostSubmission } from "../../../apiManager/services/FormServices";
+import { push } from "connected-react-router";
+import { toast } from "react-toastify";
+import { getProcessReq } from "../../../apiManager/services/bpmServices";
+import { applicationCreate } from "../../../apiManager/services/applicationServices";
 const BundleSubmit = () => {
-  const dispatch = useDispatch();
-  const isAuthenticated = useSelector((state) => state.user.isAuthenticated);
-  const tenantKey = useSelector((state) => state.tenants?.tenantId);
-  const redirectUrl = MULTITENANCY_ENABLED ? `/tenant/${tenantKey}/` : "/";
-  const bundleData = useSelector((state) => state.process.formProcessList);
-  const selectedForms = useSelector((state) => state.bundle.selectedForms);
-  const loading = useSelector((state) => state.bundle.bundleLoading);
-  const userRoles = useSelector((state)=> state.user.roles);
-  const isDesigner = userRoles.includes(STAFF_DESIGNER);
-  const { error } = useSelector((state) => state.form);
-  const [validationErrorIndex , setValidationErroIndex] = useState(null);
-  const formSubmissionError = useSelector(
-    (state) => state.formDelete.formSubmissionError
-  );
-  const [formStep, setFormStep] = useState({step:0});
-  const [getFormLoading, setGetFormLoading] = useState(false);
-  const [form, setForm] = useState({}); 
-  const bundleSubmission = useSelector(
-    (state) => state.bundle.bundleSubmission
-  );
-  const [submission, setSubmission] = useState(null);
-  const [bundleSubmitLoading, setBundleSubmitLoading] = useState(false);
-  const formRef = useRef();
-  const [formCache, setFormCache] = useState({});
-  let formValidationNotOver = true; 
-  useEffect(() => {
-    getForm();
-  }, [formStep]);
-
-  useEffect(() => {
-    // get fist form in initial stage
-    if (!submission && selectedForms.length) {
-      getForm();
-    }
-  }, [selectedForms]);
-
-  useEffect(()=>{
-    return()=>{
-    dispatch(setBundleSelectedForms([]));
-    };
-  },[]);
-
-  const getForm = () => {
-    if (selectedForms?.length) {
-      dispatch(clearFormError("form"));
-      setGetFormLoading(true);
-      fetchFormById(selectedForms[formStep.step].formId)
-        .then((res) => {
-          !formCache[res.data._id] &&
-            setFormCache({ ...formCache, [res.data._id]: res.data });
-          setForm(res.data);
-        })
-        .catch((err) => {
-          dispatch(
-            setFormFailureErrorData("form", err.response?.data || err.message)
-          );
-        })
-        .finally(() => {
-          setGetFormLoading(false);
-        });
-    }
-  };
-
-  const handleSubmisionData = () => {
+ 
+    const { bundleId } = useParams();
+    const dispatch = useDispatch();
+    const loading = useSelector((state) => state.bundle.bundleLoading);
+    const userRoles = useSelector((state)=> state.user.roles);
+    const isDesigner = userRoles.includes(STAFF_DESIGNER);
+    const isAuthenticated = useSelector((state) => state.user.isAuthenticated);
+    const tenantKey = useSelector((state) => state.tenants?.tenantId);
+    const redirectUrl = MULTITENANCY_ENABLED ? `/tenant/${tenantKey}/` : "/";
+    const bundleData = useSelector((state) => state.process.formProcessList);
+    const selectedForms = useSelector((state) => state.bundle.selectedForms);
+    const { error } = useSelector((state) => state.form);
+    const formSubmissionError = useSelector(
+      (state) => state.formDelete.formSubmissionError
+    );
+    useEffect(()=>{
+    dispatch(setBundleLoading(true));
+    dispatch(setBundleSubmissionData({}));
+    dispatch(clearFormError("form"));
+    dispatch(clearSubmissionError("submission"));
     dispatch(
-      setBundleSubmissionData({
-        data: { ...bundleSubmission.data, ...submission.data },
+      getFormProcesses(bundleId, (err, data) => {
+        if (err) { 
+          dispatch(setFormFailureErrorData("form", err));
+          dispatch(setBundleLoading(false));
+        } else {
+          executeRule({},data.id)
+          .then((res) => {
+            dispatch(setBundleSelectedForms(res.data));
+           })
+          .catch((err) => {
+           dispatch(setFormFailureErrorData("form", err));
+          })
+          .finally(() => {
+            dispatch(setBundleLoading(false));
+          });
+        }
       })
     );
+    return()=>{
+      dispatch(setBundleSelectedForms([]));
+      };
+  },[bundleId, dispatch]);
+
+  const onConfirmSubmissionError = () => {
+    const ErrorDetails = {
+      modalOpen: false,
+      message: "",
+    };
+    dispatch(setFormSubmissionError(ErrorDetails));
   };
 
-  const checkFormStepChange = (responseData)=>{
-    let changed = null;
-    if (responseData) {
-      responseData.forEach((i,index)=> { 
-        if(changed === null && selectedForms[index]?.formId !== i.formId){
-          changed = index;
-        }
-      });
-    }
-    return changed;
-  };
-
-  const handleNextForm = () => {
-    handleSubmisionData();
-    if (formRef.current.formio.checkValidity()) {
-      if(validationErrorIndex !== null){
-        setValidationErroIndex(null);
-      }
-      setBundleSubmitLoading(true);
-      executeRule({ data: submission.data }, bundleData.id)
-        .then((res) => {
-          let changed = null;
-          if(  res.data.length - 1 !== formStep.step){
-             changed = checkFormStepChange(res.data); 
-          }
-          dispatch(setBundleSelectedForms(res.data));
-          setFormStep({step: changed ? changed : formStep.step + 1});
+  const onSubmit = (bundleSubmission, bundleId) =>{
+    formioPostSubmission(bundleSubmission, bundleId, true)
+        .then((res) => { 
+          const origin = `${window.location.origin}${redirectUrl}`;
+          const data = getProcessReq({_id: bundleData.formId}, res.data._id, origin);
+          dispatch(applicationCreate(data),(err)=>{
+            if(err){
+              toast.error("Application not created");
+            }else{
+              toast.success("Submission Saved.");
+              dispatch(push(`${redirectUrl}${isDesigner ? 'bundle' : 'form'}`));
+            }
+          });
          
-        })
-        .finally(() => {
-          setBundleSubmitLoading(false);
-        });
-    }
-  };
-
-  const handleBackForm = () => {
-    setFormStep({step:formStep.step - 1});
-  };
-
-  const bundleFormValidation = async (valid, index) => {
-    if (!valid && formValidationNotOver) {
-      formValidationNotOver = valid;
-      setValidationErroIndex(index);
-      setFormStep({step:index});
-      setBundleSubmitLoading(false);
-      return;
-    }
-    if (valid && index === selectedForms.length - 1 && formValidationNotOver) {
-      formValidationNotOver = false;
-      const response = await executeRule(
-        { data: submission.data },
-        bundleData.id
-      );
-      if (response && response.data.length - 1 !== formStep.step) {
-        const changed = checkFormStepChange(response.data); 
-        dispatch(setBundleSelectedForms(response.data));
-        setFormStep({step:changed});
-        setBundleSubmitLoading(false);
-        return;
-      }
-
-      formioPostSubmission(bundleSubmission, bundleData.formId, true)
-        .then(() => {
-          toast.success("Submission Saved.");
-          dispatch(push(`${redirectUrl}${isDesigner ? 'bundle' : 'form'}`));
         })
         .catch(() => {
           const ErrorDetails = {
@@ -171,35 +92,8 @@ const BundleSubmit = () => {
           dispatch(setFormSubmissionError(ErrorDetails));
         })
         .finally(() => {
-          setBundleSubmitLoading(false);
+           dispatch(setBundleSubmitLoading(false));
         });
-    }
-  };
-
-  const handleSubmit = async () => {
-    handleSubmisionData();
-    if (formRef.current.formio.checkValidity()) {
-      setBundleSubmitLoading(true);
-      selectedForms.forEach(async (form, index) => {
-        const formioForm = await Formio.createForm(formCache[form.formId]);
-        formioForm.submission = { data: submission.data };
-        formioForm.on("change", function (changed) {
-          const { isValid } = changed;
-          formioForm.setPristine(false);
-          formioForm.checkValidity();
-          formioForm.redraw();
-          bundleFormValidation(isValid, index);
-        });
-      });
-    }
-  };
-
-  const onConfirmSubmissionError = () => {
-    const ErrorDetails = {
-      modalOpen: false,
-      message: "",
-    };
-    dispatch(setFormSubmissionError(ErrorDetails));
   };
 
   if (loading) {
@@ -236,68 +130,7 @@ const BundleSubmit = () => {
         <div>
           {!selectedForms.length ? <Errors errors={error} /> : ""}
           {selectedForms.length ? (
-            <LoadingOverlay
-              active={bundleSubmitLoading}
-              spinner
-              text={"Loading..."}
-            >
-              <div className="border py-2">
-                <Stepper activeStep={formStep.step} nonLinear alternativeLabel>
-                  {selectedForms?.map((form,index) => (
-                    <Step key={form.id}>
-                      <StepLabel error={validationErrorIndex === index}>{form.formName}</StepLabel>
-                    </Step>
-                  ))}
-                </Stepper>
-                <div className="px-3">
-                  {getFormLoading ? (
-                    <Loading />
-                  ) : (
-                    <div>
-                      <Errors errors={error} />
-                      <h3 className="px-3 py-2">{form.title}</h3>
-                      <div className="px-3 py-2">
-                        <Form
-                          form={form}
-                          options={{ hide: { submit: true } }}
-                          ref={formRef}
-                          submission={bundleSubmission}
-                          onChange={(e) => {
-                            setSubmission(e);
-                          }}
-                        />
-                      </div>
-
-                      <div className="d-flex align-items-center justify-content-end px-3 py-2">
-                        {formStep.step === 0 ? (
-                          ""
-                        ) : (
-                          <button
-                            onClick={handleBackForm}
-                            className="btn btn-secondary mr-2"
-                          >
-                            Previous Form
-                          </button>
-                        )}
-                        <button
-                          onClick={
-                            selectedForms.length - 1 === formStep.step
-                              ? handleSubmit
-                              : handleNextForm
-                          }
-                          disabled={bundleSubmitLoading}
-                          className="btn btn-primary"
-                        >
-                          {selectedForms.length - 1 === formStep.step
-                            ? "Submit Form"
-                            : "Next Form"}
-                        </button>
-                      </div>
-                    </div>
-                  )}
-                </div>
-              </div>
-            </LoadingOverlay>
+           <BundleSubmission onSubmit={onSubmit}/>
           ) : (
             <h3 className="text-center">No Forms Selected</h3>
           )}
